requirements.txt
# Core web frameworks and servers
fastapi
starlette
uvicorn[standard]
httpx

# MCP (Model Context Protocol)
fastmcp # <--- CHANGE THIS FROM mcpx TO fastmcp

# Agent Framework
langchain-mcp-adapters
langgraph

# Langchain and LLM integration (core components only, no specific integrations)
langchain
langchain-community
langchain-core
langchain-groq

# Environment management
python-dotenv